Added fetch_with_retry method and asyncio library
A new `fetch_with_retry` method has been introduced to enable asynchronous fetch with retry option. This method handles connection errors by waiting 2 seconds before retrying for `ClientConnectionError`. In addition, it handles a timeout error and retries the download up to three times. Diagnostic information is output to the console. This new functionality is intended to increase the reliability of the resource download process.