Create the parse_html method
A refactoring of the `parse_html` method has been done, where the BeautifulSoup library is used with the "lxml" parser. A ready-made BS4 structure for the parsed HTML content is now returned. This modification is intended to make the code more readable and provide a more consistent interface for HTML processing.