Update of the get_number_of_pages function
The `get_number_of_pages` function has been refactored, improving the process of retrieving and analysing the number of pages. The function now uses the `fetch_and_parse` function to retrieve and parse HTML content. The way the number of pages is obtained through more precise CSS selectors has also been updated. In addition, it has been optimised to update the progress bar value only when it is available.